requirements.txt
cassandra-driver==3.5.0
Django==4.0.3
django-cassandra-engine==1.7.0
pytest==7.1.1

PyTumblr==0.1.0
beautifulsoup4==4.8.1
Pillow==9.1.0
setuptools==44.0.0
wheel==0.37.1
asgiref==3.5.0
django-widget-tweaks==1.4.8
djangorestframework==3.12.4
image==1.5.33

pytz==2021.1
six==1.15.0
sqlparse==0.4.1

requests==2.27.1

